# Coalesced memory read for a slightly faster LLM interence