## Background
I enjoyed reading the posts related to this tweet: https://twitter.com/awnihannun/status/1776275621926375498
I'd like to contribute to the MLX side in the challenge with Llama.cpp, even if it's just a small improvement. :)

## Proposed changes
This PR modifies the way `scales` and `biases` are read from memory, making the access more coalesced.
Although the data in these variables are sparsely arranged for a SIMD thread, they can be read in a more adjacent manner when considering all SIMD threads together.

## Performance Benchmarks
Performance improved around `0.2%` with this change.

- Machine: M3MAX (128GB)
- LLM: mistral-7b (4Q) 
- Client: https://github.com/ml-explore/mlx-examples
- Command: `python -m mlx_lm.generate --model mlx_model --prompt "write a quicksort algorithm in c++" --max-tokens 256`
- Before: main(tag: v0.9.1)
- After: this PR

Token Processing Speed (tokens per second, higher is better, 15 runs):
|       | Before | After | Speedup |
|-------|--------|-------|---------|
| Median| 68.554 | 68.680| 0.18%   |
| Mean  | 68.557 | 68.734| 0.26%   |
| Min   | 68.210 | 68.592| 0.56%   |
| Max   | 68.800 | 69.087| 0.42%   |

## Checklist
- [x] I have read the [CONTRIBUTING](https://github.com/ml-explore/mlx/blob/main/CONTRIBUTING.md) document
- [x] I have run `pre-commit run --all-files` to format my code and installed pre-commit prior to committing changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have updated the necessary documentation (if needed)

## Additional Notes
- All tests passed.
- Applying `clang-format` to `quantized.metal` resulted in significant formatting changes. To avoid overwhelming diffs, I opted to bypass the linting step for this file.