@youknow04 2024-04-10
Thank you for your detailed feedback and insights. Let me address each of your points:

- Tiny performance gain issue
Matrix-vector multiplication is a fundamental operation and the core computational load in Transformers.
Optimizing this primitive operation can indeed have a persistent effect on all variants.

- Faster for loop issue
First, I should mention that this is only my second time programming in Metal, so I don't have expertise in the Metal framework itself.
However, I have done such low-latency optimizations many times, and unrolling is not always faster than a for loop, especially when the unrolled code is too long.
The code itself should be in memory, so larger code size implies more memory overhead, such as cache misses.
But I didn't expect the for loop to be faster than the unrolled version here.
I appreciate your intensive checking.

- Experiment issue
I don't think it's a throttling issue because I manually tested each run with checks.
I tested performance with GPU monitoring, ran the LLM when GPU utilization was idle, executed each run with enough rest to prevent throttling issues, and the cooling fan was not triggered in my M3MAX (128G) for all tests.

I think we now need the LLN(Law of Large Numbers) to test this tiny gain in the LLM.
MLX is still in its early stages, and you must have a lot of work to do.
I want to support the MLX team and really don't want to take your time for this hacky gain.
I will make the test code for intensive test and share the result again.