@youknow04 2024-04-14
I tried to reproduce the results using the following steps:

- reboot my mac
- run`sudo sysctl iogpu.wired_lwm_mb=100000`
- run following python script
    - this is randomized script to prevent possible biases.

<details>
<summary>Used python code to reproduce</summary>
```python
import json
import os
import random
import statistics
import subprocess
import time

from dataclasses import dataclass

MLX_LM_PATH = os.getcwd()
MLX_PATH = os.path.expanduser("~/workspace/mlx")
NUM_BATCH = 32
NUM_ITER = 999
RESULT_FILE = "result.jsonl"

@dataclass
class Target:
    branch: str
    commit_hash: str

targets = [
    Target("main", "bddf23f175726a57f0e443cd45518c0757daa166"),
    Target("coalesced-mem", "ab58e3718a3e31687e3ef5e8914034855990454f"),
    Target("coalesced-mem-unroll-t", "28057ab228de7d1b1042622559b4a1ef7ba14a12"),
    # Target("coalesced-mem-for-1", "893a6327b7ec1d2b391c6e8e27822fecbf5d0e04"),
    # Target("coalesced-mem-for-2", "af9a0269a35ce68409e4f4f5e09d6f82da55c835"),
]

def run_command(command: list[str]):
    return subprocess.check_output(command).decode("utf-8")


def setup_mlx(target: Target):
    os.chdir(MLX_PATH)
    run_command(["git", "switch", target.branch])

    print(f"installing mlx[{target.branch}]")
    run_command(["pip", "install", "."])
    pip_list_result = subprocess.check_output(["pip", "list"]).decode("utf-8")
    if f"+{target.commit_hash[:8]}" not in pip_list_result:
        raise RuntimeError(f"wrong MLX version installed. {target.branch}")
    os.chdir(MLX_LM_PATH)

def show_result():
    parsed: dict[str, list[float]] = {t.branch: [] for t in targets}
    with open(RESULT_FILE, "r") as f:
        for line in f:
            result = json.loads(line)
            for r in result:
                parsed[r["branch"]].append(float(r["tps"]))

    for branch, tps_values in parsed.items():
        mean_tps = statistics.mean(tps_values)
        median_tps = statistics.median(tps_values)
        print(f"{branch}: Mean = {mean_tps}, Median = {median_tps} Min = {min(tps_values)}, Max = {max(tps_values)}")


if __name__ == "__main__":
    show_result()   
    exit()  
    os.chdir(MLX_LM_PATH)

    for i in range(NUM_ITER):
        r_samples = random.sample(targets, len(targets))
        print(f"{i}th iteration with {r_samples}")
        sample_result: list[dict[str, str|float]] = []
        for target in r_samples:
            setup_mlx(target)
            for b in range(NUM_BATCH):
                time.sleep(2)  # to prevent throttling
                llm_result = run_command(["python", "-m", "mlx_lm.generate", "--model", "mlx_mistral7b-8q", "--prompt", "'write a quicksort algorithm in c++'", "--max-tokens", "128"])
                tps = float(llm_result.split("Generation: ")[1].split(" tokens-per-sec")[0])
                sample_result.append({
                    "branch": target.branch,
                    "iter": i,
                    "batch": b,
                    "tps": tps,
                })
        with open(RESULT_FILE, "a") as f:
            f.write(json.dumps(sample_result)+"\n")
        show_result()
```
</details>

and got this result

- `main`: v0.9.1
- `coalesced-mem`: this pr
- `coalesced-mem-unroll-t`: this PR with https://github.com/ml-explore/mlx/pull/803
- `coalesced-mem-for-1`: this PR using one for loop 
- `coalesced-mem-for-2`: this PR using tow for loop 

<img width="587" alt="mistral7b-4bit" src="https://github.com/ml-explore/mlx/assets/7802242/f77ace77-4b01-40a0-91d8-49214eb4d192">
<img width="575" alt="mistral7b-8bit" src="https://github.com/ml-explore/mlx/assets/7802242/bee346fd-1789-4fb9-ac2a-d863d762ec8f">

# Result Summary:
- After `rebooting` and the `iogpu.wired_lwm_mb` setting, all inference speeds increased, and the speed gap between different configurations was reduced.
- The rank order remained consistent.
- The combination of this PR and https://github.com/ml-explore/mlx/pull/803 shows the best performance.
- We now have clean results, and throttling or other operational issues have minimal impact due to randomization and the LLN.
- This PR delivers improved performance, and it is even faster when combined with https://github.com/ml-explore/mlx/pull/803.

In my opinion, for truly low-latency code, complex implementations are inevitable.
we also have room for improvement when we choose this direction.

@angeloskath let me know what you think.